Fix application issues and prepare for deployment
This commit addresses several issues that were preventing the application from running and being deployed.

- Creates a `.env` file to provide the necessary `VITE_ADMIN_EMAIL` environment variable.
- Removes hardcoded admin credentials from the authentication context for improved security.
- Refactors `App.jsx` to remove a redundant `AuthProvider` and moves the `ProtectedRoute` component to its own file for better code organization.
- Adds a `vercel.json` file to configure rewrites for single-page application routing on Vercel.
