+++ .env
@@ -1,0 +1,1 @@
+VITE_ADMIN_EMAIL=admin@example.com
